Use a uint64_t in computing the tiles permutations.

+++ src/tiles/tiles_state.cc
@@ -22,15 +22,15 @@ void TilesState::compute_hash(void)
 {
 	unsigned int bits = 0;
 	const Tiles *t = dynamic_cast<const Tiles *>(domain);
-	const vector<unsigned int> *ones = t->get_ones();
+	const vector<uint64_t> *ones = t->get_ones();
 	const vector<uint64_t> *fact_ary = t->get_fact_ary();
 
 	hash_val = 0;
 	for (int i = tiles.size() - 1; i >= 0; i -= 1) {
-		unsigned int k = tiles[i];
-		unsigned int mask = ~((~0) << k);
-		unsigned int v = mask & bits;
-		unsigned int d = k - ones->at(v);
+		uint64_t k = tiles[i];
+		uint64_t mask = ~((~0) << k);
+		uint64_t v = mask & bits;
+		uint64_t d = k - ones->at(v);
 		hash_val += d * fact_ary->at(i);
 		bits |= 1 << k;
 	}
